dai0kou-backend/functions/generate_task/requirements.txt
functions-framework==3.*
google-events==0.4.0
google-api-core==2.13.0
protobuf==4.22.1
cloudevents==1.9.0
google-cloud-tasks==2.13.1